# doc/api.yml
openapi: 3.0.3
info:
  title: Diploma API
  description: >
    Gestion des cours, notes (historisees), relevés PDF (S3 + email async) et export XLSX
    des diplomes pour HEI (parcours EL / TN).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://ja3rfnp2d6pc2nxspnbniaqpvm0mhlho.lambda-url.eu-west-3.on.aws/
    description: Preprod
security:
  - basicAuth: []
...
paths:
...
  /students/{studentId}/group:
    post:
      tags: [Groups]
      summary: Changer le groupe d'un etudiant (ADMIN)
      description: >
        Ferme l'appartenance courante (to_date=now) et en ouvre une nouvelle.
        Les notes de l'etudiant ne sont jamais impactees.
      parameters:
        - $ref: '#/components/parameters/StudentIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeGroupRequest' }
      responses:
        '201':
          description: Nouvelle entree d'historique de groupe creee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentGroupHistory' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /students/{studentId}/group-history:
    get:
      tags: [Groups]
      summary: Historique des groupes d'un etudiant
      description: L'etudiant ne peut consulter que le sien ; TEACHER/ADMIN consultent librement.
      parameters:
        - $ref: '#/components/parameters/StudentIdPathParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StudentGroupHistory' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /students/{studentId}/grades:
    get:
      tags: [Grades]
      summary: Lister les notes d'un etudiant
      description: L'etudiant ne voit que les siennes ; TEACHER/ADMIN voient celles de tous.
      parameters:
        - $ref: '#/components/parameters/StudentIdPathParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Grade' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Grades]
      summary: Poser ou corriger une note (TEACHER assigne au cours, ou ADMIN)
      description: >
        La raison est obligatoire des qu'une note existe deja pour cet (etudiant, examen) ;
        chaque changement est historise (voir GradeHistory).
      parameters:
        - $ref: '#/components/parameters/StudentIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetGradeRequest' }
      responses:
        '201':
          description: Note posee/corrigee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grade' }
        '400':
          description: Raison obligatoire pour corriger une note existante
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
...
    StudentIdPathParam:
      name: studentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Acces refuse
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    Track:
      type: string
      enum: [EL, TN]
    Role:
      type: string
      enum: [STUDENT, TEACHER, ADMIN]
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
    AppUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
    Student:
      type: object
      properties:
        id: { type: string, format: uuid }
        user: { $ref: '#/components/schemas/AppUser' }
        firstName: { type: string }
        lastName: { type: string }
        track: { $ref: '#/components/schemas/Track' }
        enrollmentYear: { type: integer, example: 2024 }
...
    Course:
      type: object
      properties:
        id: { type: string, format: uuid }
        ref: { type: string, example: ALG101 }
        title: { type: string }
        credits: { type: integer }
        track:
          allOf: [{ $ref: '#/components/schemas/Track' }]
          nullable: true
          description: null = tronc commun (EL et TN)
    Group:
      type: object
      properties:
        id: { type: string, format: uuid }
        ref: { type: string, example: EL1-A }
        track: { $ref: '#/components/schemas/Track' }
        academicYear: { type: integer, example: 2024 }
    StudentGroupHistory:
      type: object
      properties:
        id: { type: string, format: uuid }
        student: { $ref: '#/components/schemas/Student' }
        group: { $ref: '#/components/schemas/Group' }
        fromDate: { type: string, format: date-time }
        toDate:
          type: string
          format: date-time
          nullable: true
          description: null = appartenance courante
...
    Exam:
      type: object
      properties:
        id: { type: string, format: uuid }
        course: { $ref: '#/components/schemas/Course' }
        title: { type: string }
        dateExam: { type: string, format: date-time }
        coefficient: { type: number, format: double, example: 0.25 }
        academicYear: { type: integer }
        semester: { type: integer, enum: [1, 2] }
    Grade:
      type: object
      properties:
        id: { type: string, format: uuid }
        student: { $ref: '#/components/schemas/Student' }
        exam: { $ref: '#/components/schemas/Exam' }
        value: { type: number, format: double, example: 14.5 }
        lastModifiedAt: { type: string, format: date-time }
...
    ChangeGroupRequest:
      type: object
      required: [newGroupId]
      properties:
        newGroupId: { type: string, format: uuid }
...
    SetGradeRequest:
      type: object
      required: [examId, value]
      properties:
        examId: { type: string, format: uuid }
        value: { type: number, format: double, example: 14.5 }
        reason:
          type: string
          nullable: true
          description: Obligatoire si une note existe deja pour cet (etudiant, examen)